Refactor module
