+++ src/penn_chime/cli.py
@@ -3,14 +3,14 @@
 import os
 import sys
 
-from .parameters import Parameters
-from .models import SimSirModel as Model
+from .model.parameters import Parameters
+from .model.sir import Sir
 
 
 def main():
     """Main."""
     p = Parameters.create(os.environ, sys.argv[1:])
-    m = Model(p)
+    m = Sir(p)
 
     for df, name in (
         (m.sim_sir_w_date_df, "sim_sir_w_date"),
@@ -18,7 +18,3 @@ def main():
         (m.census_df, "projected_census"),
     ):
         df.to_csv(f"{p.current_date}_{name}.csv")
-
-
-if __name__ == "__main__":
-    main()
