+++ src/penn_chime/view/charts.py
@@ -1,13 +1,10 @@
-from datetime import datetime
-from math import ceil
 from typing import Dict, Optional
 
 from altair import Chart
 import pandas as pd
 import numpy as np
 
-from .constants import DATE_FORMAT
-from .parameters import Parameters
+from ..constants import DATE_FORMAT
 
 
 def build_admits_chart(
