+++ st_app.py
@@ -1,0 +1,13 @@
+"""
+Run the streamlit app.
+
+`pip install .` is required.
+
+Module is in site-packages, and does not reflect any modifications made in the src directory.
+"""
+
+from penn_chime.view.st_app import main
+
+
+if __name__ == '__main__':
+    main()
